zaloha

+++ preprocess.py
@@ -4,8 +4,13 @@ import json
 import numpy as np
 from tqdm import tqdm
 import cv2
-from api.tools import Tools
+from api.tools import Tools as t
 from names import Names as n
+import time,os, cv2, json
+import numpy as np
+from tqdm import tqdm
+from api.tools import Tools as t
+from timeout import timeout
 
 class CreateProcessPool:
 
@@ -21,28 +26,30 @@ class CreateProcessPool:
 
 
     def run(self):
-        self.yt_list = self.load_json(self.youtube_path)
-        self.ex_list = self.load_json(self.exports_path)
-        self.youtube_to_process, self.export_to_process  = {},{}
+
+
+        self.yt_list = t.load_json(self.youtube_path)
+        self.ex_list = t.load_json(self.exports_path)
+        self.youtube_to_process, self.export_to_process  = [],[]
+        print('==> prepare_in_out_list')
         self.prepare_in_out()
 
+
+
+
+        self.youtube_to_process_path = os.path.join(self.workdir, 'youtube_to_process.json')
+        self.export_to_process_path = os.path.join(self.workdir, 'export_to_process.json')
+
+        print('==> read youtube durations')
         self.read_durations(self.youtube_to_process)
+        t.save_json(self.youtube_to_process_path, self.youtube_to_process)
+
+        print('==> read export durations')
         self.read_durations(self.export_to_process)
+        t.save_json(self.export_to_process_path, self.export_to_process)
 
+        CreateImages().process(self.yt_list, self.ex_list, self.workdir)
 
-        self.save_json(os.path.join(self.workdir, 'youtube_to_process.json'), self.youtube_to_process)
-        self.save_json(os.path.join(self.workdir, 'export_to_process.json'), self.export_to_process)
-
-    @staticmethod
-    def load_json(path):
-        with open(path, 'r') as f:
-            data  = json.load(f)
-        return data
-
-    @staticmethod
-    def save_json(path, data):
-        with open(path, 'w', encoding='utf8') as outfile:
-            json.dump(data, outfile, indent=4, ensure_ascii=False)
 
     @staticmethod
     def record(src, dst):
@@ -58,72 +65,85 @@ class CreateProcessPool:
 
             name, _ = os.path.splitext(splitted[-1])
             out_path = os.path.join(n.EXPORT, event_id, name)
-            self.export_to_process[out_path] = self.record(path, out_path)
+            self.export_to_process.append(self.record(path, out_path))
 
         for path in self.yt_list:
             basename = os.path.basename(path)
             name, _ = os.path.splitext(basename)
             out_path = os.path.join(n.YOUTUBE, name)
-            self.youtube_to_process[out_path] = self.record(path, out_path)
+            self.youtube_to_process.append(self.record(path, out_path))
 
 
 
 
     def read_durations(self, data):
-        for key, record in tqdm(data.items(), ascii=True, desc='process {} records'.format(len(data))):
+        for idx in tqdm(range(len(data)), ascii=True, desc='ffprobe on {} files'.format(len(data))):
+            record = data[idx]
             try:
-                duration = Tools.get_duration(record[n.SRC])
-                data[key][n.DURATION] = duration
+                duration = t.get_duration(record[n.SRC])
+                data[idx][n.DURATION] = duration * 1000
             except Exception as e:
                 print(record, e)
 
 
+class CreateImages:
+    def __init__(self):
+        pass
+    def process(self, yt_path, ex_path, workdir):
 
-    def process(self):
-        # Afilename, _ = os.path.splitext(self.main_list_path)
-        # Bfilename, _ = os.path.splitext(self.second_list_path)
+        yt_list = t.load_json(yt_path)
+        ex_list = t.load_json(ex_path)
+        self.test_len = 0
+        self.run_from = 41609
+        self.timestamp=  time.time()
+        self.list_to_process = yt_list + ex_list
 
-        process_list_a = []
-        for idx, (in_path, out_path) in enumerate(self.list_to_process):
 
-            if self.test_len !=0 and idx > self.test_len:
-                continue
+
+        self.err_file = os.path.join(workdir, 'errors.json')
+
 
         run_to = self.run_from + self.test_len if self.test_len != 0 else len(self.list_to_process)
+
 
         errors = []
         for idx in tqdm(range(run_to), ascii=True, desc='process {} links'.format(run_to)):
             if idx < self.run_from:
                 continue
-            in_path, out_path = self.list_to_process
-            name = '{}.png'.format(self.timestamp)
-            name_info = '{}.info'.format(self.timestamp)
+            record = self.list_to_process[idx]
+            path, out_path_rel = record['src'], record['dst']
+            out_path = t.create_folder(os.path.join(workdir, out_path_rel))
+
+
+
+            name = 'image.png'
+            name_info = 'info.json'
             out = os.path.join(out_path, name)
             out_info = os.path.join(out_path, name_info)
+            if os.path.isfile(out):
+                continue
 
             try:
                 # print(f_idx)
                 frame, info = self.get_full_frame(path)
+                info['src']=path
                 # np.save(out, frame)
                 cv2.imwrite(out, frame)
-                with open(out_inf, 'w') as outfile:
+                with open(out_info, 'w') as outfile:
                     json.dump(info, outfile, indent=4)
 
             except Exception as e:
-                err=(f_idx, path, e)
+                err = '{}: {}'.format(path, e)
                 print(err)
                 errors.append(err)
 
             # print(info)
         if len(errors):
-
             with open(self.err_file, 'w') as outfile:
                 json.dump(errors, outfile, indent=4)
 
-
-
+    @timeout(10)
     def get_full_frame(self, path):
-
         cap = cv2.VideoCapture(path)
         fps = cap.get(cv2.CAP_PROP_FPS)
         frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
@@ -136,7 +156,7 @@ class CreateProcessPool:
             ret, frame = cap.read()
             rets.append(ret)
             if not ret:
-                frame = np.zeros((360,640,3), dtype=np.uint8)
+                frame = np.zeros((360, 640, 3), dtype=np.uint8)
             else:
                 frame = cv2.resize(frame, (640, 360))
 
@@ -146,13 +166,13 @@ class CreateProcessPool:
                 'timecodes': np.ndarray.tolist(timecodes),
                 'fps': '{}'.format(fps),
                 'frame_count': '{}'.format(frame_count),
-                'rets':rets,
-                'timestamp':self.timestamp
+                'rets': rets,
+
+                'timestamp': self.timestamp
                 }
         a = np.concatenate((frames[0:3]), axis=0)
         b = np.concatenate((frames[3:6]), axis=0)
         c = np.concatenate((frames[6:9]), axis=0)
-
 
         full_frame = np.concatenate((a, b, c), axis=1)
         #
@@ -163,29 +183,12 @@ class CreateProcessPool:
 
         return full_frame, info
 
-    def run_list(self, video_list, img_dir=None, get_images=False):
-        dur_len = len(video_list) if self.test_len is None else self.test_len
-        durations = np.zeros(dur_len, dtype=np.int)
-        for idx in tqdm(range(dur_len), ascii=True, desc='checks {} links'.format(dur_len)):
-            path = video_list[idx]
-            cap = cv2.VideoCapture(path)
-            try:
-                fps = cap.get(cv2.CAP_PROP_FPS)
-                frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-                duration = int(1000*(frame_count/fps))
+# '''/mnt/data/palpatine/DATASETS/YT_LINK/workdir/export_to_process_no_dur.json
+# /mnt/data/palpatine/DATASETS/YT_LINK/workdir/youtube_to_process_no_dur.json'''
+# yt_list = '/mnt/data/palpatine/SAMPLES/YT_LINK/REF/workdir/youtube_to_process.json'
+# ex_list = '/mnt/data/palpatine/SAMPLES/YT_LINK/REF/workdir/export_to_process.json'
+# workdir = '/mnt/data/palpatine/SAMPLES/YT_LINK/REF/workdir'
 
-            except:
-                duration = 0
-            durations[idx] = duration
-
-            cap.release()
-        return durations
-
-
-class Match:
-    def __init__(self, ida):
-        self.ida = ida
-        self.idb = []
-        self.diff = []
-        self.adur = None
-        self.bdur = None
+# yt_list = '/mnt/data/palpatine/DATASETS/YT_LINK/workdir/youtube_to_process_no_dur.json'
+# ex_list = '/mnt/data/palpatine/DATASETS/YT_LINK/workdir/export_to_process_no_dur.json'
+# workdir = '/mnt/data/palpatine/DATASETS/YT_LINK/workdir'
